+++ google/cloud/spanner_v1/database.py
@@ -1227,6 +1227,7 @@ class BatchSnapshot(object):
         partition_size_bytes=None,
         max_partitions=None,
         data_boost_enabled=False,
+        directed_read_options=None,
         *,
         retry=gapic_v1.method.DEFAULT,
         timeout=gapic_v1.method.DEFAULT,
@@ -1266,6 +1267,12 @@ class BatchSnapshot(object):
                 (Optional) If this is for a partitioned read and this field is
                 set ``true``, the request will be executed via offline access.
 
+        :type directed_read_options: :class:`~google.cloud.spanner_v1.types.DirectedReadOptions`
+            or :class:`dict`
+        :param directed_read_options: (Optional) Request level option used to set the directed_read_options
+            for ReadRequests that indicates which replicas
+            or regions should be used for non-transactional reads.
+
         :type retry: :class:`~google.api_core.retry.Retry`
         :param retry: (Optional) The retry settings for this request.
 
@@ -1294,6 +1301,7 @@ class BatchSnapshot(object):
             "keyset": keyset._to_dict(),
             "index": index,
             "data_boost_enabled": data_boost_enabled,
+            "directed_read_options": directed_read_options,
         }
         for partition in partitions:
             yield {"partition": partition, "read": read_info.copy()}
@@ -1338,6 +1346,7 @@ class BatchSnapshot(object):
         max_partitions=None,
         query_options=None,
         data_boost_enabled=False,
+        directed_read_options=None,
         *,
         retry=gapic_v1.method.DEFAULT,
         timeout=gapic_v1.method.DEFAULT,
@@ -1389,6 +1398,12 @@ class BatchSnapshot(object):
                 (Optional) If this is for a partitioned query and this field is
                 set ``true``, the request will be executed via offline access.
 
+        :type directed_read_options: :class:`~google.cloud.spanner_v1.types.DirectedReadOptions`
+            or :class:`dict`
+        :param directed_read_options: (Optional) Request level option used to set the directed_read_options
+            for ExecuteSqlRequests that indicates which replicas
+            or regions should be used for non-transactional queries.
+
         :type retry: :class:`~google.api_core.retry.Retry`
         :param retry: (Optional) The retry settings for this request.
 
@@ -1413,6 +1428,7 @@ class BatchSnapshot(object):
         query_info = {
             "sql": sql,
             "data_boost_enabled": data_boost_enabled,
+            "directed_read_options": directed_read_options,
         }
         if params:
             query_info["params"] = params
